Refactor dice system
